pyproject.toml
[project]
name = "iikoserver-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "iikoserver-client",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20250915",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv.sources]
iikoserver-client = { git = "https://github.com/UserVanya/Pub-iikoserver-py-sdk.git" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require real API access)",
    "slow: Slow tests",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
